netty: Status should be based on GOAWAY code
goingAway() is called before onGoAwayRead() in Netty:
https://github.com/netty/netty/blob/b7f57223c14b495afb11f1ce0fe32b3d21185a08/codec-http2/src/main/java/io/netty/handler/codec/http2/DefaultHttp2ConnectionDecoder.java#L521

The test before checked that the stream went away, but not that the
GOAWAY code influenced our Status, as UNAVAILABLE is the default
internally.

The UNAVAILABLE default has also been changed to include a message so
that we can determine where the Status came from in case it is triggered
again in the future.