De-flake transport test.
See #289 for more detail.

I can't reproduce the failure locally, so my guess is that our TIMEOUT is too short for Travis, when the Travis machines get high CPU usage or short of memory (and causes full GC), we should tolerate longer waiting time.

I'm not sure whether this change can fix the flakeness, but I'd like to give it a try.